+++ lib/quickbooks_web_connector/errors.rb
@@ -1,0 +1,6 @@
+module QuickbooksWebConnector
+
+  class RequestXMLError < StandardError
+  end
+
+end