+++ lib/quickbooks_web_connector/soap_wrapper/defaultServant.rb
@@ -74,10 +74,20 @@ module QuickbooksWebConnector
       #   parameters      SendRequestXMLResponse - {http://developer.intuit.com/}sendRequestXMLResponse
       #
       def sendRequestXML(parameters)
-        job = QuickbooksWebConnector::Job.peek
-        request_xml = job ? job.request_xml : nil
-
-        SendRequestXMLResponse.new request_xml
+        if (job = QuickbooksWebConnector::Job.peek)
+          case (request_xml = job.request_xml)
+          when :failed
+            raise RequestXMLError
+          else
+            SendRequestXMLResponse.new request_xml
+          end
+        else
+          SendRequestXMLResponse.new nil
+        end
+      rescue RequestXMLError
+        # Remove the job from the queue since it fails to build. The job should have already created a failure.
+        QuickbooksWebConnector.pop
+        retry
       end
 
       # SYNOPSIS